BACKEND/requirements.txt
asgiref==3.8.1
Django>=4.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt==5.2.2
drf-yasg
django-cors-headers
setuptools
djangorestframework-simplejwt
django-filter
python-dotenv